sdk/java: read TLS CA from environment
This change is intended to make it easier to write one
Java program that works both on production and inside
with-ledgerd. It adds one new environment variable to
control the Java SDK's behavior:

- SEQTLSCA specifies a path to a CA certificate file.
It replaces ca_file in config param ssl_params.

This removes the equivalent configuration methods on the
builder class. We are approaching a new major version (v2)
anyway, so that's okay.

Closes #3103

Author: Keith Rarick <kr@xph.us>
Date: Thu Mar 29 15:08:02 2018 -0700
upstream:5cddd15e6b51e7f6eaf8aede41e0188cb7aaf7a8