sdk/java: upgrade to okhttp3
Prepping the way for Java 9/10/11 support with okhttp3,
as well as a gson upgrade to 2.8.2 as a dependency.

We should consider another 1.5.x release to deprecate
the `setConnectTimeout` and other methods on the
`Client` object that are no longer viable with the
builder pattern.

Closes #3437

Author: Dominic Dagradi <ddagradi@gmail.com>
Date: Mon Apr 23 13:17:20 2018 -0700
upstream:7147e5a520adef1a76b6c78e7efd005eed9ed728