sdk/java: refresh session based on TTL
If we ever change the hostname for a ledger,
we don't want client software to need to be restarted
to pick up the new value.

The `/hello` endpoint returns a TTL for this purpose.
When the TTL expires, re-build the HTTP client
used for all API requests in the SDK.

If there is a problem with subsequent `/hello` requests,
it is okay for the SDK to continue using the address it already knows
while it continues trying for a successful `/hello`.
However, if the SDK has a problem making its initial `/hello` request,
it should report an error to the application.

Closes #4120

Author: Dan Croak <dan@statusok.com>
Date: Fri Jul 13 15:28:44 2018 -0700
upstream:6ac65927ea57f12f32fe31c0fdb9bc0894aabcd3